+++ pkg/util/listUtils.go
@@ -3,11 +3,11 @@ package util
 func GetDuplicates[T comparable](list []T) []T {
 	elementCount := CountGrouped(list)
 
-	// get list of names with count != 1
+	// get list of values with count != 1
 	var duplicates []T
-	for name, count := range elementCount {
+	for value, count := range elementCount {
 		if count != 1 {
-			duplicates = append(duplicates, name)
+			duplicates = append(duplicates, value)
 		}
 	}
 	return duplicates