+++ CHANGELOG.md
@@ -20,9 +20,6 @@ General:
 - [#15] Check component health
 - [#17] add health checks before and after applying the blueprint
 - [#22] Add `dryRun` option. If `dryRun` is active the blueprint procedure stops before applying resources to the cluster and remains in the actual state. One can set the option to false and continue at this state.
-- [#36] Set registry configuration for dogu and global config.
-- [#45] Set registry configuration for encrypted values.
-
 
 Dogu-specific:
 - [#4] introduce flag `allowDoguNamespaceSwitch` for dogu namespace switch
@@ -37,3 +34,6 @@ Component-specific:
 
 Config-specific:
 - [#42] Implement config repositories
+- [#48] Save config diff in cluster
+- [#36] Set registry configuration for dogu and global config.
+- [#45] Set registry configuration for encrypted values.